requirements.txt
tensorflow
python-chess
Flask
numpy
tqdm